+++ source/Calamari.AzureWebApp/Calamari.AzureWebApp.csproj
@@ -23,4 +23,8 @@
     <ProjectReference Include="..\Calamari.Azure\Calamari.Azure.csproj" />
     <ProjectReference Include="..\Calamari.CloudAccounts\Calamari.CloudAccounts.csproj" />
   </ItemGroup>
+  
+  <ItemGroup Condition=" '$(TargetFramework)' == 'net462' ">
+    <PackageReference Include="Microsoft.Web.Deployment" Version="4.0.5"/>
+  </ItemGroup>
 </Project>
