+++ source/Calamari.AzureWebApp/Integration/Websites/Publishing/WebDeployPublishSettings.cs
@@ -1,6 +1,6 @@
 ﻿namespace Calamari.AzureWebApp.Integration.Websites.Publishing
 {
-    class WebDeployPublishSettings {
+    public class WebDeployPublishSettings {
         /// <summary>
         /// The deployment site name may be different for web deploy especially when using slots. This may
         /// end up as something like slotname___sitename and is what needs to be provided to web deploy.
